Finished first working version of httpusher
