+++ Cargo.toml
@@ -1,14 +1,21 @@
 [package]
-name = "beam-httpusher"
+name = "samply_beam_httpusher"
 version = "0.1.0"
 edition = "2021"
+license = "Apache-2.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[target.x86_64-unknown-linux-gnu]
-rustflags = [
-    "-C", "link-arg=-fuse-ld=lld",
-]
+[profile.release]
+#opt-level = "z"     # Optimize for size.
+lto = true          # Enable Link Time Optimization
+codegen-units = 1   # Reduce number of codegen units to increase optimizations.
+panic = "abort"     # Abort on panic
+strip = true        # Automatically strip symbols from the binary.
+
+[profile.bloat]
+inherits = "release"
+strip = false
 
 [dependencies]
 shared = { path = "../beam/shared" }
